chore: separate the hook
